# config/openapi.yaml
openapi: 3.0.0
info: 
  title: Community Events API
  description: event management api
  version: 1.0.0
  contact:
    name: Author
    email: stephen@enterity.com
servers:
  - url: https://dev.community-events.com/v1
    description: Development server
security:
  - JWT: []
components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: app.auth.decode_token
...
  responses:
...
    Ok:
      description: Ok
  schemas:
...
    Comment:
      properties:
        id:
          type: string
          format: uuid
        contents:
          type: string
        resolved:
          type: boolean
        edited:
          type: boolean
    Reply:
      properties:
        id:
          type: string
          format: uuid
        contents:
          type: string
        resolved:
          type: boolean
        edited:
          type: boolean
        parent_id:
          type: string
          format: uuid        
...
paths:
...
  /events/{id}/comments:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: app.events.add_comment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
              - $ref: '#/components/schemas/Comment'
              - $ref: '#/components/schemas/Reply'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
    get:
      operationId: app.events.get_comments
      responses:
        '200':
          description: Ok
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
...